#73 - re-fix types, fix headers
