+++ src/hooks/useCommands.tsx
@@ -1,6 +1,5 @@
-import VisibilityOffIcon from '@mui/icons-material/VisibilityOff'
-import { Box, IconButton, Tooltip } from '@mui/material'
 import useAxios from 'axios-hooks'
+import { generateCommandName } from 'components/generateCommandName'
 import { ServerConfigContainer } from 'containers/ConfigContainer'
 import { MakeItHappenButton } from 'pages/CommandIndex'
 import {
@@ -52,21 +51,6 @@ export const useCommands = () => {
     includeHidden,
     hiddenOnChange,
   }
-}
-
-export const generateCommandName = (hidden: boolean, name: string) => {
-  return hidden ? (
-    <Tooltip title="hidden command" aria-label={name + ' (hidden)'}>
-      <Box component="span">
-        {name}
-        <IconButton disabled>
-          <VisibilityOffIcon color="disabled" fontSize="small" />
-        </IconButton>
-      </Box>
-    </Tooltip>
-  ) : (
-    name
-  )
 }
 
 const commandMapper = (command: Command): CommandRow => {
