+++ src/pages/CommandBlocklistView/CommandBlocklistView.tsx
@@ -2,11 +2,12 @@ import AddIcon from '@mui/icons-material/Add'
 import DeleteIcon from '@mui/icons-material/Delete'
 import { Box, Button, Checkbox, IconButton, Tooltip } from '@mui/material'
 import Divider from 'components/divider'
+import { generateCommandName } from 'components/generateCommandName'
 import { ModalWrapper } from 'components/ModalWrapper'
 import PageHeader from 'components/PageHeader'
 import { Table } from 'components/Table'
 import { useBlockList } from 'hooks/useBlockList'
-import { generateCommandName, useCommands } from 'hooks/useCommands'
+import { useCommands } from 'hooks/useCommands'
 import { useModalColumns, useTableColumns } from 'pages/CommandBlocklistView'
 import { useMemo, useState } from 'react'
 import { CommandBase, CommandRow } from 'types/custom_types'
