+++ src/pages/CommandBlocklistView/commandBlocklistViewHelpers.tsx
@@ -8,21 +8,27 @@ const useModalColumns = () => {
       {
         Header: 'Add',
         accessor: 'action',
-        width: 85,
+        disableSortBy: true,
+        disableGroupBy: true,
+        disableFilters: true,
+        canHide: false,
+        width: 60,
       },
       {
         Header: 'Namespace',
         accessor: 'namespace',
-        filter: 'fuzzyText',
+        width: 150,
       },
       {
         Header: 'System',
         accessor: 'system',
         filter: 'fuzzyText',
+        width: 150,
       },
       {
         Header: 'Command',
         accessor: 'name',
+        width: 300,
       },
     ],
     [],
@@ -35,25 +41,31 @@ const useTableColumns = () => {
       {
         Header: 'Namespace',
         accessor: 'namespace',
-        filter: 'fuzzyText',
+        width: 170,
       },
       {
         Header: 'System',
         accessor: 'system',
         filter: 'fuzzyText',
+        width: 170,
       },
       {
         Header: 'Command',
         accessor: 'name',
+        width: 320,
       },
       {
         Header: 'Status',
         accessor: 'status',
-        filter: 'fuzzyText',
+        width: 170,
       },
       {
         Header: '',
         accessor: 'action',
+        disableSortBy: true,
+        disableGroupBy: true,
+        disableFilters: true,
+        canHide: false,
         width: 85,
       },
     ],
